extract cache into separate module away from the port
