+++ committer/src/setup.rs
@@ -9,7 +9,7 @@ use metrics::{
 };
 use services::{
     block_committer::{port::l1::Contract, service::BlockCommitter},
-    fee_metrics_tracker::{port::cache::CachingApi, service::FeeMetricsTracker},
+    fee_metrics_tracker::{cache::CachingApi, service::FeeMetricsTracker},
     state_committer::port::Storage,
     state_listener::service::StateListener,
     state_pruner::service::StatePruner,
