+++ fee_algo_simulation/src/main.rs
@@ -2,7 +2,7 @@ use std::net::SocketAddr;
 
 use actix_web::web::{self, Data};
 use anyhow::Result;
-use services::fee_metrics_tracker::port::cache::CachingApi;
+use services::fee_metrics_tracker::cache::CachingApi;
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::EnvFilter;
